Implement checks for proofs of registration
get_address and sign_psbt now fail if the provided proof of
registration is missing or not valid

Special cases (e.g. single-sig) where the PoR is not needed are
not part of this and will be added in the future.

Also completes the implementation of the corresponding feature in
the bitcoin V-App client and CLI interface.