Replace COMM_REF to a struct containing callbacks
This allows fully decoupling the Nbgl code from the Comm type, by
using a static object with callbacks that the Comm object can fill
in the constructor. Alternative implementations of the Comm object
can therefore provide equivalent hooks via the same callbacks.

Therefore, the init_comm function is now empty, as the callbacks
are filled in the constructor of the Comm object, that should be a
singleton.