[DRAFT] Implement hmac for repeated pages
Incomplete: using a placeholder instead of the actual HMAC
(that should use a private key, possibly key deterministically
crated from the app hash (this would allow stored HMACs to
remain valid for future executions of the app).

Exploratory commit to modify the protocol to compute an hmac
before dropping a code page, so that the required data for pages
sent multiple times is reduced; moreover, the client could (in
principle) store it permanently for future executions of the
V-App).