Add a `test-mode` feature to app-sdk to avoid expecting a client
When running unit tests in a V-App, using any `println!` would
hang because in native execution, that expects streaming prints
via the TCP connection.
The `test-mode` feature in the app-sdk replaces this behaviour
of prints by just printing on the console.
This allows running unit tests that cover code that contains
calls to the print! or println! macro.